backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.2
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
redis==5.0.1
Pillow==10.0.0
python-dotenv==1.0.0
pymysql==1.1.0
cryptography==41.0.3
Werkzeug==2.3.7
bcrypt==4.0.1
email-validator==2.0.0.post2
alembic==1.12.0
gunicorn==21.2.0
pytest==7.4.0
PyJWT==2.8.0
requests==2.31.0 